refactor(python): organise PyBoy modules into a pyboy/ subpackage
Moves pyboy.py, pyboy_api.py, pyboy_wrappers.py and pyboy_debug.py
into a boytacean/pyboy/ subpackage as core.py, api.py, wrappers.py
and debug.py respectively. Adds boytacean/pyboy/__init__.py that
re-exports the public surface so `from boytacean.pyboy import PyBoy`
keeps working unchanged for external consumers.

The new layout mirrors upstream PyBoy's own structure
(pyboy.api / pyboy.plugins) and makes the architectural split
between the base GameBoy wrapper and the PyBoy compatibility shim
explicit. All 44 pytest tests still pass.
